+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased - xxxx-xx-xx
 
+### Removed
+
+- Removed the deprecated `FragmentArguments` derive.
+- Removed the deprecated `arugment_struct` attribute.
+
 ## v3.0.0-beta.3 - 2023-05-13
 
 ### New Features